% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{keys}
\alias{key_field_domain_id}
\alias{key_age_range}
\alias{relevel_age}
\alias{key_plots_by_type}
\alias{key_spp}
\title{Key functions for NEON data}
\usage{
key_field_domain_id(id)

key_age_range(age_range, current_year = 2023)

relevel_age(age_range)

key_plots_by_type(plot_id)

key_spp(code_vec)
}
\description{
Sometimes you need to key abbreviations and such!
}
