% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_functions}
\alias{plot_functions}
\alias{HighResTiff}
\alias{lm_eqn}
\alias{Multiplot}
\alias{NMDS_screeplot}
\alias{c25}
\alias{plot_vp}
\alias{ess_plot}
\alias{param_plot}
\alias{plot_resid}
\alias{pp_check}
\title{Plot functions for NEON1 analysis}
\usage{
HighResTiff(plot_obj, file, width_in, height_in, resolution_dpi, ...)

lm_eqn(df0, xvar, yvar)

Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)

NMDS_screeplot(mat, kmax = 10)

c25(option = 1)

plot_vp(hM, VP)

ess_plot(es_df, var, ylab)

param_plot(post, ylab)

plot_resid(m_p, mp_p)

pp_check(m_p, mp_pp, n_draws = NULL)
}
\description{
NA
}
