% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{external_functions}
\alias{external_functions}
\alias{left_join}
\title{External functions}
\usage{
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = NULL
)
}
\description{
NA
}
