% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{logit_Warton}
\alias{check_residuals}
\alias{standardize_df}
\title{Helper functions for NEON1}
\usage{
logit_Warton(y, epsilon = NULL)

check_residuals(fit, dir, grp = NULL, nm = NULL)

standardize_df(df_in)
}
\description{
Self-explanatory
}
